Extract Schwab option reconciliation into its own module
schwab.py had grown by 440 lines of option lifecycle handling on top of the
CSV reading it already does. The lifecycle moves to schwab_options.py
unchanged, and reconciliation splits into a scan that reads the rows as
lifecycles and an emission that replaces the terminal rows, so neither is
longer than one screen. No behaviour change: the option tests pass untouched.
